covPCA {pcaPP}	R Documentation

Robust Covariance Matrix Estimation

Description

computes the robust covariance matrix using the PCAgrid and PCAproj functions.

Usage

covPCAproj(x, control)
covPCAgrid(x, control)

Arguments

x
a numeric matrix or data frame which provides the data.

control
a list whose elements must be the same as (or a subset of) the parameters of the appropriate PCA function (PCAgrid or PCAproj).

Details

The functions covPCAproj and covPCAgrid use the functions PCAproj and PCAgrid respectively to estimate the covariance matrix of the data matrix x.

Value

cov
the actual covariance matrix estimated from x

center
the center of the data x that was substracted from them before the PCA algorithms were run.

method
a string describing the method that was used to calculate the covariance matrix estimation

Author(s)

Heinrich Fritz, Peter Filzmoser <P.Filzmoser@tuwien.ac.at>

References

C. Croux, P. Filzmoser, M. Oliveira, (2007). Algorithms for Projection-Pursuit Robust Principal Component Analysis, Chemometrics and Intelligent Laboratory Systems, Vol. 87, pp. 218-225.

See Also

PCAgrid, ScaleAdv, princomp

Examples

  # multivariate data with outliers
  library(mvtnorm)
  x <- rbind(rmvnorm(200, rep(0, 6), diag(c(5, rep(1,5)))),
             rmvnorm( 15, c(0, rep(20, 5)), diag(rep(1, 6))))
  covPCAproj(x)
  # compare with classical covariance matrix:
  cov(x)

[Package pcaPP version 2.0-5 Index]